Refactor: Hit interface 분리
