+++ src/main/java/game/Element.java
@@ -1,5 +1,9 @@
 package game;
 
+import game.hit.Hit;
+import game.hit.HitBall;
+import game.hit.HitNone;
+import game.hit.HitStrike;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -9,10 +13,16 @@ public class Element<T> {
     private Integer position;
     private T value;
 
-    public boolean isPositionMatched(Element<T> other){
+    public Hit beHit(Element<T> other){
+        if(!this.isValueMatched(other)) return new HitNone();
+        if(!this.isPositionMatched(other)) return new HitBall();
+        return new HitStrike();
+    }
+
+    private boolean isPositionMatched(Element<T> other){
         return position.equals(other.position);
     }
-    public boolean isValueMatched(Element<T> other){
+    private boolean isValueMatched(Element<T> other){
         return value.equals(other.value);
     }
 }
