+++ src/main/java/game/hit/Hit.java
@@ -1,0 +1,7 @@
+package game.hit;
+
+import game.Result;
+
+public interface Hit {
+    void update(Result result);
+}
