+++ src/main/java/game/hit/HitBall.java
@@ -1,0 +1,10 @@
+package game.hit;
+
+import game.Result;
+
+public class HitBall implements Hit{
+    @Override
+    public void update(Result result) {
+        result.increaseBalls();
+    }
+}
