+++ src/main/java/game/hit/HitNone.java
@@ -1,0 +1,10 @@
+package game.hit;
+
+import game.Result;
+
+public class HitNone implements Hit{
+    @Override
+    public void update(Result result) {
+        
+    }
+}
