+++ src/main/java/game/hit/HitStrike.java
@@ -1,0 +1,10 @@
+package game.hit;
+
+import game.Result;
+
+public class HitStrike implements Hit{
+    @Override
+    public void update(Result result) {
+        result.increaseStrikes();
+    }
+}
